Added suggestions for Module 2. Added Unit and Integration tests.
