+++ .env.j2
@@ -1,0 +1,9 @@
+# Flask server configuration.
+FLASK_APP=todo_app/app
+FLASK_ENV=production
+
+# Change the following values for local development.
+TRELLO_API_KEY={{ trello_apy_key }}
+TRELLO_API_TOKEN={{ trello_token }}
+TRELLO_URL={{ trello_url }}
+TRELLO_BOARD_ID={{ trello_board_id }}
