+++ ansible/deploy_todoApp.yaml
@@ -1,0 +1,51 @@
+---
+- name: Deploy To Do app
+  hosts: webserver
+  remote_user: ec2-user
+
+  tasks: 
+    - name: Install Git
+      yum:
+        name: git
+        state: present
+      become: true
+    
+    - name: Install Python 3
+      yum:
+        name: python3
+        state: present
+      become: true
+
+    - name: Install Poetry
+      ansible.builtin.shell: curl -sSL https://install.python-poetry.org | python3 -
+      args:
+        creates: /home/ec2-user/.local/bin/poetry
+
+    - name: Create directory for the To-Do App code
+      ansible.builtin.file:
+        path: /opt/todoapp
+        state: directory
+        mode: '0755'
+        owner: ec2-user
+      become: true
+
+    - name: Get the latest version of your code with Git
+      ansible.builtin.git:
+        repo: 'https://github.com/CreaTiVeUK/DevOps-Course-Starter.git'
+        dest: /opt/todoapp
+        version: module-3
+    
+    - name: Install project dependencies
+      ansible.builtin.shell: source ~/.bash_profile && poetry install
+      args:
+        chdir: /opt/todoapp
+
+    - name: Create a .env file on the host
+      ansible.builtin.template:
+      src: /mytemplates/foo.j2
+      dest: /etc/file.conf
+      mode: '0644'
+      vars:
+        
+
+
