+++ ansible/inventory.txt
@@ -1,0 +1,5 @@
+[webserver]
+18.135.68.84
+
+#"ControllerIP" = "35.177.90.75"
+#"ManagedNodeIP" = "18.135.68.84"
