+++ todo_app/app.py
@@ -1,6 +1,5 @@
 from flask import Flask, render_template, request, url_for, redirect
-from todo_app.flask_config import Config
-from todo_app.data.item import Item
+from todo_app.data.configs.flask_config import Config
 from todo_app.data.ViewModel import ViewModel
 import todo_app.data.session_items_trello as session
 import todo_app.data.configs.logger_settings as logger
@@ -21,7 +20,7 @@ def create_app():
     def add_item():
         title = request.form['filmTitle']
         logger.log.info('Title to add: ' + title)
-        Item.create_Card('To Do', title)
+        session.create_Card('To Do', title)
         return redirect(url_for('index'))
 
 
@@ -30,7 +29,7 @@ def create_app():
         description = request.form['description']
         logger.log.info('Card ID\'s description to update: ' + titleID)
         logger.log.info('Description: ' + description)
-        Item.update_item_description(titleID, description)
+        session.update_item_description(titleID, description)
         return redirect(url_for('index'))
 
 
@@ -39,31 +38,29 @@ def create_app():
         due = request.form['date']
         logger.log.info('Card ID\'s : ' + titleID)
         logger.log.info('Card due date: ' + due)
-        Item.update_item_due_date(titleID, due)
+        session.update_item_due_date(titleID, due)
         return redirect(url_for('index'))
 
 
     @app.route('/complete_item/<string:titleID>', methods=['POST'])
     def complete_item(titleID):
-        logger.log.info('Attempting to move card')
-        logger.log.info('Item to move from To Do to Done: ' + titleID)
-        Item.move_Card('Done', titleID)
+        logger.log.info('Attempting to move card to Done list')
+        session.move_Card('Done', titleID)
         return redirect(url_for('index'))
 
 
     @app.route('/doing_item/<string:titleID>', methods=['POST'])
     def doing_item(titleID):
-        logger.log.info('Attempting to move card')
-        logger.log.info('Item to move from To Do to Done: ' + titleID)
-        Item.move_Card('Done', titleID)
+        logger.log.info('Attempting to move card to Doing list')
+        session.move_Card('Doing', titleID)
         return redirect(url_for('index'))
 
 
     @app.route('/revert_item/<string:titleID>', methods=['POST'])
     def revert_item(titleID):
-        logger.log.info('Attempting to move card')
-        logger.log.info('Item to revert from Done to To Do: ' + titleID)
-        Item.move_Card('To Do', titleID)
+        logger.log.info('Attempting to move card to To Do list')
+        session.move_Card('To Do', titleID)
         return redirect(url_for('index'))
     
     return app
+
